commands: make symlinks for iroffer
Having iroffer watch DATA_DIR caused it to release the anime on IRC
earlier than it was scheduled - right after the upload script finished
transferring anime files. This was bad. I configured iroffer it to use a
separate watchdir just for the IRC releases. This means that new
releases need to be symlinked into this directory to get published on
IRC, but this commit automatizes this process. The outcome is that
scheduled releases now will happen at the same hour both on
nyaa.si/anidex as well as IRC.